:hammer: Refactor analytical solution of nodal coordinates to elements
