+++ CMakeLists.txt
@@ -93,6 +93,7 @@ endif()
 # Include directories
 include_directories(BEFORE
   ${mpm_SOURCE_DIR}/include/
+  ${mpm_SOURCE_DIR}/include/elements/
   ${mpm_SOURCE_DIR}/external/
   ${mpm_SOURCE_DIR}/tests/include/
 )
@@ -126,8 +127,13 @@ if(MPM_BUILD_TESTING)
     ${mpm_SOURCE_DIR}/tests/cell_test.cc
     ${mpm_SOURCE_DIR}/tests/factory_test.cc
     ${mpm_SOURCE_DIR}/tests/geometry_test.cc
-    ${mpm_SOURCE_DIR}/tests/hexahedron_element_test.cc
-    ${mpm_SOURCE_DIR}/tests/hexahedron_quadrature_test.cc  
+    ${mpm_SOURCE_DIR}/tests/elements/hexahedron_element_test.cc
+    ${mpm_SOURCE_DIR}/tests/elements/hexahedron_quadrature_test.cc  
+    ${mpm_SOURCE_DIR}/tests/elements/quadrilateral_element_test.cc
+    ${mpm_SOURCE_DIR}/tests/elements/quadrilateral_gimp_element_test.cc
+    ${mpm_SOURCE_DIR}/tests/elements/quadrilateral_quadrature_test.cc    
+    ${mpm_SOURCE_DIR}/tests/elements/triangle_element_test.cc
+    ${mpm_SOURCE_DIR}/tests/elements/triangle_quadrature_test.cc
     ${mpm_SOURCE_DIR}/tests/io_test.cc
     ${mpm_SOURCE_DIR}/tests/material/bingham_test.cc
     ${mpm_SOURCE_DIR}/tests/material/linear_elastic_test.cc
@@ -146,12 +152,7 @@ if(MPM_BUILD_TESTING)
     ${mpm_SOURCE_DIR}/tests/particle_container_test.cc
     ${mpm_SOURCE_DIR}/tests/particle_test.cc
     ${mpm_SOURCE_DIR}/tests/point_in_cell_test.cc
-    ${mpm_SOURCE_DIR}/tests/quadrilateral_element_test.cc
-    ${mpm_SOURCE_DIR}/tests/quadrilateral_gimp_element_test.cc
-    ${mpm_SOURCE_DIR}/tests/quadrilateral_quadrature_test.cc    
     ${mpm_SOURCE_DIR}/tests/read_mesh_ascii_test.cc
-    ${mpm_SOURCE_DIR}/tests/triangle_element_test.cc
-    ${mpm_SOURCE_DIR}/tests/triangle_quadrature_test.cc
     ${mpm_SOURCE_DIR}/tests/write_mesh_particles.cc
     ${mpm_SOURCE_DIR}/tests/write_mesh_particles_unitcell.cc
   )
