+++ include/elements/triangle_element.h
@@ -1,126 +1,60 @@
-#ifndef MPM_HEXAHEDRON_ELEMENT_H_
-#define MPM_HEXAHEDRON_ELEMENT_H_
+#ifndef MPM_TRIANGLE_ELEMENT_H_
+#define MPM_TRIANGLE_ELEMENT_H_
 
 #include "element.h"
 #include "logger.h"
 
-//! MPM namespace
 namespace mpm {
 
-//! Hexahedron element class derived from Element class
-//! \brief Hexahedron element
-//! \details 8-noded and 20-noded hexahedron element \n
-//! Shape function, gradient shape function, B-matrix, indices \n
-//! 8-node (Trilinear) Hexahedron Element \n
+//! Triangle element class derived from Element class
+//! \brief Triangle element
+//! \details 3-noded and 6-noded triangle element \n
+//! Shapte function, gradient shape function, B-matrix, indices \n
+//! 3-node Triangle ELement \n
 //! <pre>
-//!        3               2
-//!          0_ _ _ _ _ _0
-//!         /|           /|
-//!        / |          / |
-//!     7 0_ |_ _ _ _ _0 6|
-//!       |  |         |  |
-//!       |  |         |  |
-//!       |  0_ _ _ _ _|_ 0
-//!       | / 0        | / 1
-//!       |/           |/
-//!       0_ _ _ _ _ _ 0
-//!     4               5
-//!
-//!
+//!   2 0
+//!     |`\
+//!     |  `\
+//!     |    `\
+//!     |      `\
+//!     |        `\
+//!   0 0----------0 1
 //! </pre>
-//! 20-node (Serendipity) Hexahedron Element \n
+//! 6-node Triangle Element
 //! <pre>
-//!        3       13          2
-//!          0_ _ _ 0 _ _ _  0
-//!          /|             / |
-//!      15 0 |         14 0  |
-//!        /  0 9         /   |
-//!     7 0_ _| _ 0 _ _ _ 0 6 0 11
-//!       |   |   19     |    |
-//!       |   |      8   |    |
-//!       | 0 0_ _ _ 0 _ |_ _ 0  1
-//!    17 0  /           0 18 /
-//!       | 0 10         |  0 12
-//!       |/             | /
-//!       0_ _ _ 0 _ _ _ 0
-//!     4        16         5
-//!
+//!   2 0
+//!     |`\
+//!     |  `\
+//!   5 0    `0 4
+//!     |      `\
+//!     |        `\
+//!   0 0-----0----0 1
+//!           3
 //! </pre>
 //!
-//! 27-node (Triquadratic) Hexahedron Element \n
-//! Check with GMSH \n
-//! <pre>
-//!          3           13             2
-//!            0_ _ _ _ _ 0 _ _ _ _ _ 0
-//!            /|                     /|
-//!           / |                    / |
-//!          /  |    24             /  |
-//!      15 0   |     0         14 0   |
-//!        /    |      20 0       /    |
-//!       /   9 0                /     0 11
-//!      /      |               /      |
-//!  7  0_ _ _ _|_ 0 _ _ _ _ _ 0 6     |
-//!     |       | 19           |       |
-//!     |  0 22 |              |   0 23|
-//!     |       |    0 26  8   |       |
-//!     |     0 0_ _ _ _ _ 0_ _|_ _ _  0  1
-//!     |      /               |      /
-//!  17 0     /    0 25        0 18  /
-//!     |    /                 |    /
-//!     |10 0         0        |   0 12
-//!     |  /         21        |  /
-//!     | /                    | /
-//!     |/                     |/
-//!     0_ _ _ _ _ 0 _ _ _ _ _ 0
-//!   4           16            5
-//!
-//!
-//! </pre>
-//!
-//! Face numbering for 8-node, 20-node and 27-node Hexaheron Element \n
-//!
-//! <pre>
-//!            Behind: F4
-//!        3      F2       2
-//!          0_ _ _ _ _ _0
-//!         /|           /|
-//!        / |          / |
-//!     7 0_ |_ _ _ _ _0 6|
-//!       |  |         |  |
-//!    F3 |  |         |  |   F1
-//!       |  0_ _ _ _ _|_ 0
-//!       | / 0        | / 1
-//!       |/     F0    |/
-//!       0_ _ _ _ _ _ 0
-//!     4               5
-//!         Front: F5
-//!
-//!
-//! Bottom face: F0, Right face: F1, Top face: F2,
-//! Left face: F3, Rear face: F4, Front face: F5
-//! </pre>
 //!
 //! \tparam Tdim Dimension
 //! \tparam Tnfunctions Number of functions
 template <unsigned Tdim, unsigned Tnfunctions>
-class HexahedronElement : public Element<Tdim> {
+class TriangleElement : public Element<Tdim> {
 
  public:
-  //! Define a vector of size dimension
+  //! Define vector of size dimension
   using VectorDim = Eigen::Matrix<double, Tdim, 1>;
 
   //! constructor with number of shape functions
-  HexahedronElement() : mpm::Element<Tdim>() {
-    static_assert(Tdim == 3, "Invalid dimension for a hexahedron element");
-    static_assert((Tnfunctions == 8 || Tnfunctions == 20),
-                  "Specified number of shape functions is not defined");
+  TriangleElement() : mpm::Element<Tdim>() {
+    static_assert(Tdim == 2, "Invalid dimension for a triangular element");
+    static_assert((Tnfunctions == 3 || Tnfunctions == 6),
+                  "Specified number of shape funcions is not defined");
+
     //! Logger
-    std::string logger = "hexahedron::<" + std::to_string(Tdim) + ", " +
+    std::string logger = "triangular::<" + std::to_string(Tdim) + ", " +
                          std::to_string(Tnfunctions) + ">";
     console_ = std::make_unique<spdlog::logger>(logger, mpm::stdout_sink);
   }
 
-  //! Return number of functions
+  //! Return number of shape functions
   unsigned nfunctions() const override { return Tnfunctions; }
 
   //! Evaluate shape functions at given local coordinates
@@ -131,11 +65,11 @@ class HexahedronElement : public Element<Tdim> {
   Eigen::VectorXd shapefn(const VectorDim& xi, const VectorDim& particle_size,
                           const VectorDim& deformation_gradient) const override;
 
-  //! Evaluate local shape functions at given local coordinates
+  //! Evaluate local shape functions at given coordinates
   //! \param[in] xi given local coordinates
   //! \param[in] particle_size Particle size
   //! \param[in] deformation_gradient Deformation gradient
-  //! \retval shapefn Shape function of a given cell
+  //! \retval shapefn_local Shape function of a given cell
   Eigen::VectorXd shapefn_local(
       const VectorDim& xi, const VectorDim& particle_size,
       const VectorDim& deformation_gradient) const override;
@@ -156,10 +90,9 @@ class HexahedronElement : public Element<Tdim> {
   //! \param[in] deformation_gradient Deformation gradient
   //! \retval jacobian Jacobian matrix
   Eigen::Matrix<double, Tdim, Tdim> jacobian(
-      const Eigen::Matrix<double, 3, 1>& xi,
-      const Eigen::MatrixXd& nodal_coordinates,
-      const Eigen::Matrix<double, 3, 1>& particle_size,
-      const Eigen::Matrix<double, 3, 1>& deformation_gradient) const override;
+      const VectorDim& xi, const Eigen::MatrixXd& nodal_coordinates,
+      const VectorDim& particle_size,
+      const VectorDim& deformation_gradient) const override;
 
   //! Compute Jacobian local
   //! \param[in] xi given local coordinates
@@ -168,10 +101,9 @@ class HexahedronElement : public Element<Tdim> {
   //! \param[in] deformation_gradient Deformation gradient
   //! \retval jacobian Jacobian matrix
   Eigen::Matrix<double, Tdim, Tdim> jacobian_local(
-      const Eigen::Matrix<double, 3, 1>& xi,
-      const Eigen::MatrixXd& nodal_coordinates,
-      const Eigen::Matrix<double, 3, 1>& particle_size,
-      const Eigen::Matrix<double, 3, 1>& deformation_gradient) const override;
+      const VectorDim& xi, const Eigen::MatrixXd& nodal_coordinates,
+      const VectorDim& particle_size,
+      const VectorDim& deformation_gradient) const override;
 
   //! Evaluate the B matrix at given local coordinates for a real cell
   //! \param[in] xi given local coordinates
@@ -227,15 +159,26 @@ class HexahedronElement : public Element<Tdim> {
   //! \retval indices Indices that make the face
   Eigen::VectorXi face_indices(unsigned face_id) const override;
 
-  //! Return the number of faces in a hexahedron
-  unsigned nfaces() const override { return 6; }
+  //! Return the number of faces in a triangle
+  unsigned nfaces() const override { return 3; }
 
   //! Return unit element length
-  double unit_element_length() const override { return 2.; }
+  double unit_element_length() const override { return 1.; }
 
   //! Return quadrature of the element
   std::shared_ptr<mpm::Quadrature<Tdim>> quadrature(
       unsigned nquadratures = 1) const override;
+
+  //! Return if natural coordinates can be evaluates
+  bool isvalid_natural_coordinates_analytical() const override;
+
+  //! Compute Natural coordinates of a point (analytical)
+  //! \param[in] nodal_coordinates Coordinates of nodes forming the cell
+  //! \param[in] point Location of the point in cell
+  //! \retval xi Return the local coordinates
+  VectorDim natural_coordinates_analytical(
+      const VectorDim& point,
+      const Eigen::MatrixXd& nodal_coordinates) const override;
 
  private:
   //! Logger
@@ -243,6 +186,6 @@ class HexahedronElement : public Element<Tdim> {
 };
 
 }  // namespace mpm
-#include "hexahedron_element.tcc"
+#include "triangle_element.tcc"
 
-#endif  // MPM_HEXAHEDRON_ELEMENT_H_
+#endif  // MPM_TRIANGLE_ELEMENT_H_
