+++ tests/elements/quadrilateral_gimp_element_test.cc
